Move DateField to flat structure
