+++ packages/components/src/components/DateField/index.ts
@@ -9,16 +9,4 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './InputField/InputField'
-export * from './InputField/InputError'
-export * from './TextInput'
-export * from './Select'
-export * from './RadioGroup/RadioGroup'
-export * from './TextArea'
-export * from './SubSectionDivider'
-export * from './VerifyingIndicator'
-export * from './ValidIndicator'
-export * from './CheckboxGroup/CheckboxGroup'
-export * from './PDFViewer/PDFViewer'
-export * from './InputField/InputLabel'
-export * from './FormTabs'
+export * from './DateField'
